refactor: Move VM into a subfolder
