+++ vm/test/casting.test.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 
-#include "vm/vm.hpp"
+#include "vm/machine/vm.hpp"
 
 using std::tuple;
