+++ vm/test/vm.test.cpp
@@ -1,4 +1,4 @@
-#include "vm/vm.hpp"
+#include "vm/machine/vm.hpp"
 
 #include <array>
